perf: improve performance of new unresolved-reference rule
We've had a few heavy rules in the past, but nothing like this
one. That's not really due to the implementation of the rule, but
simply the fact that it aggregates every ref found everywhere, which
will be a lot no matter what.

Using the standard "regal linting itself" benchmark on main before the
rule was merged and after shows a 10%+ increase in memory allocated as
well as wall clock time:

```
1111071791 ns/op	3001557280 B/op	58184205 allocs/op
1289814875 ns/op	3363166528 B/op	64788506 allocs/op - unresolved-reference merged
```

There are some design issues to address with aggregate rules that likely
would mean bigger wins across the board, but the focus of this PR is solely
to optimize the code for the rule itself.

While this remains an expensive rule, it's now closer to our other more costly rules.
In total, this saves about 2.7 million allocations of the 6.7 million allocs increased
with its addition. The steps in which this was done are roughly as follows:

```
1249527041 ns/op	3337275400 B/op	64261924 allocs/op - exclude input refs
1247609750 ns/op	3325728208 B/op	64061112 allocs/op - get rid of _import_aliases
1264200584 ns/op	3324900160 B/op	64037543 allocs/op - simplify _all_full_path_refs
1182358875 ns/op	3322581056 B/op	64013278 allocs/op - inline parts of _is_resolved_ref
1189268416 ns/op	3317109520 B/op	63942894 allocs/op - skip if name in ast.rule_and_function_names
1182895291 ns/op	3305238432 B/op	63739044 allocs/op - only aggregate data used in report
1179427375 ns/op	3273346464 B/op	63188538 allocs/op - custom result.location alternative
1161155834 ns/op	3210160768 B/op	62267845 allocs/op - defer to_location_object to report
1130275167 ns/op	3204268280 B/op	62150928 allocs/op - a few small tweaks
```

I have a few more ideas for improvements, but as they'd touch on more than just this rule,
this is probably a good place to stop for now.

Signed-off-by: Anders Eknert <anders@styra.com>