Enable prefer-values-in-head (#1535)
Although with `only-scalars` set to `true`. I did review the violations
reported without that, and while most didn't warrant "fixing", a a few
did, and so I did.

Also added an option to the rule to be able to list variable names that
should be excluded.

It's fun to play around, but the high-level goal here is mostly to have
`regal lint bundle` run all rules, even if with some exceptions defined
in the config.

Signed-off-by: Anders Eknert <anders@styra.com>