Apply modernize fixes
Not a whole lot of them, but the contains checks are nice! I wish this would
integrate with golangci-lint, but for now I don't think the overhead of running
this separately in CI is worth it, as I don't expect us to have many things to
fix in future code.

```
go run golang.org/x/tools/gopls/internal/analysis/modernize/cmd/modernize@latest -fix ./...
```

Also, a handful of style fixes.

Signed-off-by: Anders Eknert <anders@styra.com>