e2e: add frankenstein test build
This is meant to use the Regal version from whatever is in the current
checkout, together with OPA as defined in Regal's go.mod.

It's an early warning against dependency problems that only become
apparent when we're using all of OPA, not the subset of its
functionality that Regal already uses.

Signed-off-by: Stephan Renatus <stephan@styra.com>