lsp: Code Action feature rewritten in Rego
This is admittedly one of the simpler features, but all the
more reason to have it ported :) What's new in this implementation
is that we now also honor the `only` attribute sent from the client.
This is an array clients can send to only request code actions of
a certain kind, like "quickfix" or "source". Good clients are going
to filter this anyway, but doing it on the server means less data
to transfer. And of course a consistent experience regardless of
client used.

This is also the first LSP feature where a schema is used for the
input! And that was quite helpful for catching issues in the policy
during development. It does pose a problem in the tests though, as
optional attributes aren't allowed currently, which requires a few
hacks to work around. But this is definitely a model to follow for
future work.

While I have deleted the Go code related to this, I decided to keep
the test for the handler, as it's good to have something to test
the whole chain. I have also tested manually in VS Code, and it
works like a charm. Some benchmarks included too, showing pretty
good numbers!

Signed-off-by: Anders Eknert <anders@styra.com>