perf: avoid JSON/map roundtrips in ast.Value transforms (#1606)
Leveraging the new capabilities for this in roast, we can avoid
a lot of allocations by mapping directly from a source to an ast.Value
rather than going via a `map[string]any`, or even JSON. For now, focus
has been on how the **input** is mapped in most places of eval.
This can be improved further, but the natural next step is of course to
eliminate roundtrips the same way when we handle **output**, i.e. the
result of evaluation. I imagine the data-heavy aggregate rules in particular
could benefit greatly from this.

This PR also cleans up some in LSP handlers related to input transformations,
and takes a small step towards harmonizing the input format for those with
the goal of eventually having a single schema to describe the common parts
of all LSP request payloads.

Perf impact is close to 2.9 million allocs less in the Regal lint
benchmark, and greatly reduced ns/op and B/op.

```
944216438 ns/op    2805943268 B/op 51659258 allocs/op - main
894474146 ns/op    2668457832 B/op 48779590 allocs/op - PR
```

Signed-off-by: Anders Eknert <anders@styra.com>