lsp: add experimental lsp package for websocket communication (#1623)
To allow using Regal's LSP via websocket, a new little shim package is added
that can be used to start an LSP on an established websockets connection.

In a playground, it could be used to create one (filesystem-less) LSP session per
websocket caller.

This commit includes a little dev setup: a bit of Go code in `build/ws/main.go`
that shows how to combine this with github.com/gorilla/websocket, and use it
from a browser client. In this case, it's not a fully-featured web editor, but a small
htmx dev tool that lets us send arbitrary LSP messages via the established ws
connection.

Run `go run .` in `build/ws/` and visit http://localhost:8787. (No JS build steps.)

Relates to #764.

Signed-off-by: Stephan Renatus <stephan@styra.com>