lsp: Present filename relative to root in errors
This makes a change so that when a file is shown in an error message, we
show the filename relative to the root of the project rather than the
full URI.

This is done by changing the name of the file used in parsing.

I have also adjusted updateParse to take Opts struct instead to make it
less verbose.

Fixes https://github.com/StyraInc/regal/issues/1635

Signed-off-by: Charlie Egan <charlie@styra.com>