test: refactor e2e test, add new regal command runner
While really important, the e2e tests have been a bit painful to work with,
much because so much of the tests have been procedural scaffolding rather than
test logic. With this refactor, I'm hoping to improve that by introducing
a more declararative approach in the form of a tiny DSL for a command runner.

The result is less than half the lines of test code, and tests that IMHO
make it quite easy to quickly understand what's being tested. I hope others
feel the same :)

Signed-off-by: Anders Eknert <anders@styra.com>