Fix missing source.explore code action
This went missing without us noticing :( Due to the input not
being what the policy expected (and even verified at compile time).
We should look into how to leverage those schemas for runtime checks,
as that could save a lot of time troubleshooting. But for now, fixing
the issue and adding a regression test for the handler.

Signed-off-by: Anders Eknert <anders@styra.com>