Provide development option to load bundle from disk
This allows live editing of Rego in the language server, and
see the result propagated to the UI right away. A very nice
development experience!

This is however very much experimental at this point, and only
by setting an `REGAL_BUNDLE_PATH` to point at the bundle path
will this be enabled. Most likely a few issues left to resolve
here, but let's take it from here!

If we start using this often, we'll also want to optimize bundle
loading too, as that happens *very* frequently. Perhaps we could
have a file watcher notify us of changes or something like that,
and only refresh the bundle when that happens. @charlieegan3 let's
talk about that some time :)

Signed-off-by: Anders Eknert <anders@styra.com>