internal/io: explicitly disable lazy loading of regal bundle
For virtually all OPA users, this will make no difference. But if you
use Regal together with OPA and bundle activation plugin(s), like EOPA,
the default will be flipped. As a consequence, the internal Regal bundle
cannot be loaded, and every linting fails.

This only affects users that

1. build something that embeds Regal as Go module
2. register extensions with OPA's bundle loader

Since de-registering such extensions isn't well-supported, it's hard to
test for this.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>